Use typescript for inbox
