+++ project/src/js/ui/inbox/index.ts
